added grid coordinates
